% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_with.R
\name{dir_list}
\alias{dir_list}
\alias{rdata_list}
\title{Read multiple (.csv, .tsv, .rdata, .rds) files from a directory into a list.}
\usage{
rdata_list(path_dir = NULL, ...)
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{...}{Arguments passed to the reader function:
\itemize{
\item \code{rdata_list()} and \code{rda_list()} read with \code{get(load(x))} (\code{...} not unused).
\item \code{rds_list()} reads with \link{FIXME}.
\item \code{csv_list()} reads with \link{FIXME}.
\item \code{delim_list()} reads with \link{FIXME}.
\item \code{tsv_list()} reads with \link{FIXME}.
}}
}
\value{
A list of dataframes.
}
\description{
These functions wrap \code{\link[=read_with]{read_with()}}. Notice the name-format \code{<input>_list()}.
}
\seealso{
\code{\link[=read_with]{read_with()}}.

Other general functions to import data: \code{\link{read_with}}
}
\concept{general functions to import data}
