% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_path.R
\name{format_path}
\alias{format_path}
\title{Format paths.}
\usage{
format_path(files, ext, base = ".", prefix = NULL)
}
\arguments{
\item{base, prefix, files, ext}{Character vectors (all of length-1 except
\code{files}, which can be of any length.}
}
\value{
A string of the same length as \code{files}. For example:
\code{base/prefix-file1.csv} and \code{base/prefix-file2.csv}.
}
\description{
This function is most useful to map the name of each element in a list to
an output path.
}
\examples{
format_path(c("file1", "file2"), "csv")

(dfs <- list_csv(tor_example("csv")))

format_path(names(dfs), "csv")

format_path(names(dfs), "csv", base = "home", prefix = "this-")
}
