% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_any.R
\name{list_any}
\alias{list_any}
\title{Import multiple files of any format from a directory into a list.}
\usage{
list_any(
  path = ".",
  .f,
  regexp = NULL,
  ignore.case = FALSE,
  invert = FALSE,
  ...
)
}
\arguments{
\item{path}{A character vector of one path. Defaults to the working
directory.}

\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression (e.g. \verb{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{If \code{TRUE} return files which do \emph{not} match}

\item{...}{Additional arguments passed to \code{.f}.}
}
\value{
A list.
}
\description{
Import multiple files of any format from a directory into a list.
}
\examples{
tor_example()

(path <- tor_example("csv"))
dir(path)

list_any(path, read.csv)

list_any(path, ~ read.csv(.x, stringsAsFactors = FALSE))

(path_mixed <- tor_example("mixed"))
dir(path_mixed)

list_any(
  path_mixed, ~ get(load(.x)),
  "[.]Rdata$",
  ignore.case = TRUE
)

list_any(
  path_mixed, ~ get(load(.x)),
  regexp = "[.]csv$",
  invert = TRUE
)
}
\seealso{
Other functions to import files into a list: 
\code{\link{list_csv}()}

Other functions to import files of any format: 
\code{\link{load_any}()}
}
\concept{functions to import files into a list}
\concept{functions to import files of any format}
