% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_csv.R
\name{list_csv}
\alias{list_csv}
\title{Export mapping each dataframe in a list to a corresponding .csv file.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
list_csv(lst, dir, prefix = NULL)
}
\arguments{
\item{lst}{A list of dataframes.}

\item{dir}{Character; the directory where the files will be saved.}

\item{prefix}{Character; a prefix to add to the file names.}
}
\description{
Export mapping each dataframe in a list to a corresponding .csv file.
}
\examples{
lst <- list(df1 = data.frame(x = 1), df2 = data.frame(x = 2))

# Saving the output to a temporary file
output <- tempdir()
list_csv(lst, output, prefix = "myfile-")

# Look inside the output directory to confirm it worked
dir(output, pattern = "myfile")

}
\concept{functions to handle multiple spreadsheets of an excel workbook}
\concept{general functions to export data}
