% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package_data.R
\name{load_package_data}
\alias{load_package_data}
\alias{enframe_package_data}
\title{Enframe datasets from an attached package or load them into an environment.}
\usage{
load_package_data(x, envir = .GlobalEnv)

enframe_package_data(x)
}
\arguments{
\item{x}{A character string giving the name of one package.}

\item{envir}{an \code{\link{environment}} or \code{NULL}.}
}
\value{
\itemize{
\item \code{enframe_package_data()} returns a \link[tibble:tibble-package]{tibble}.`
\item \code{load_package_data()} returns invisible \code{x}.
}
}
\description{
Enframe datasets from an attached package or load them into an environment.
}
\examples{
# Must be attached
library("MASS")

enframe_package_data("MASS")

e <- new.env()
load_package_data("MASS", envir = e)
ls(e)
str(e$snails)
}
