% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_with.R
\name{read_with}
\alias{read_with}
\title{Import multiple files into a list using any given reading function.}
\usage{
read_with(.f, regexp = NULL)
}
\arguments{
\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
}
\value{
A modified version of the input function with the following
arguments:
\itemize{
\item `path_dir: String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).
\item \code{...} Arguments passed to the reader function.
}
}
\description{
Import multiple files into a list using any given reading function.
}
\examples{
readw_example()

(csv_files <- readw_example("csv"))
dir(csv_files)

csv_list <- read_with(read.csv)
csv_list(csv_files)

# Same
read_with(read.csv)(csv_files)

# More robust
csv_list <- read_with(read.csv, regexp = "[.]csv")
csv_list(csv_files, stringsAsFactors = TRUE)

# Compare
class(csv_list(csv_files)[[2]]$y)
class(csv_list(csv_files, stringsAsFactors = FALSE)[[2]]$y)

(rdata_files <- readw_example("rdata"))
dir(rdata_files)

# You may create your own reader function
read_rdata <- function(x) get(load(x))
rdata_list <- read_with(read_rdata)
rdata_list(rdata_files)
}
\concept{general functions to import data}
