% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readw_example.R
\name{readw_example}
\alias{readw_example}
\title{Easily access example data.}
\usage{
readw_example(path = NULL)
}
\arguments{
\item{path}{Length-1 character vector. A path to an available directory.}
}
\value{
A list of available directories or a path to them.
}
\description{
Easily access example data.
}
\examples{
readw_example()
readw_example("csv")
dir(readw_example("csv"))
}
\author{
Copied from \code{readr::readr_example()}, by Jim Hester and colleagues.
}
