% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwith-package.R
\docType{package}
\name{readwith-package}
\alias{readwith}
\alias{readwith-package}
\title{readwith: Read Multiple Files With Of Any Format From a Directory Into a List}
\description{
Read multiple files of any format from a directory
    into a list, using a reader function of your choice. readwith is
    designed to be simple, flexible, and to fit well in data-science
    workflows with the tidyverse.  It is focused, flexible, small, and has
    few dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/readwith}
  \item Report bugs at \url{https://github.com/maurolepore/readwith/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
\keyword{internal}
