% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwith_example.R
\name{readwith_example}
\alias{readwith_example}
\title{Easily access example data.}
\usage{
readwith_example(path = NULL)
}
\arguments{
\item{path}{Length-1 character vector. A path to an available directory.}
}
\value{
A list of available directories or a path to them.
}
\description{
Easily access example data.
}
\examples{
readwith_example()
readwith_example("csv")
dir(readwith_example("csv"))
}
\author{
Copied from \code{readr::readr_example()}, by Jim Hester and colleagues.
}
