% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_with.R
\name{write_with}
\alias{write_with}
\title{Read multiple files into a list using any given reading function.}
\usage{
write_with(.x, path = ".", .f, ext, prefix = NULL, ...)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{.f}{A function able to read the desired file format.}

\item{...}{Additional arguments passed to \link{grep}.}
}
\value{
A list.
}
\description{
Read multiple files into a list using any given reading function.
}
\examples{
tor_example()

path <- tor_example("csv")
dir(path)

list_any(path, read.csv)

list_any(path, ~read.csv(.x, stringsAsFactors = FALSE))

(path_mixed <- tor_example("mixed"))
dir(path_mixed)

list_any(
  path_mixed, ~get(load(.x)),
  regexp = "[.]csv$",
  invert = TRUE
)

list_any(
  path_mixed, ~get(load(.x)),
  "[.]Rdata$",
  ignore.case = TRUE
)
}
\seealso{
Other general functions to import data: \code{\link{list_any}},
  \code{\link{list_rds}}
}
\concept{general functions to import data}
