% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_csv.R
\name{list_csv}
\alias{list_csv}
\alias{list_tsv}
\alias{list_rds}
\alias{list_rdata}
\title{Read multiple files from a directory into a list.}
\usage{
list_csv(path = ".", regexp = "[.]csv$", ignore.case = TRUE,
  invert = FALSE, header = TRUE, sep = ",", quote = "\\"",
  dec = ".", fill = TRUE, comment.char = "",
  stringsAsFactors = FALSE, na.strings = c("", "NA"), ...)

list_tsv(path = ".", regexp = "[.]tsv$", ignore.case = TRUE,
  invert = FALSE, header = TRUE, sep = "\\t", quote = "\\"",
  dec = ".", fill = TRUE, comment.char = "",
  stringsAsFactors = FALSE, na.strings = c("", "NA"), ...)

list_rds(path = ".", regexp = "[.]rds$", ignore.case = TRUE,
  invert = FALSE)

list_rdata(path = ".", regexp = "[.]rdata$|[.]rda$",
  ignore.case = TRUE, invert = FALSE)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{If \code{TRUE} return files which do \emph{not} match}

\item{header}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}

\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{quote}{the set of quoting characters. To disable quoting
    altogether, use \code{quote = ""}.  See \code{\link{scan}} for the
    behaviour on quotes embedded in quotes.  Quoting is only considered
    for columns read as character, which is all of them unless
    \code{colClasses} is specified.}

\item{dec}{the character used in the file for decimal points.}

\item{fill}{logical. If \code{TRUE} then in case the rows have unequal
    length, blank fields are implicitly added.  See \sQuote{Details}.}

\item{comment.char}{character: a character vector of length one
    containing a single character or an empty string.  Use \code{""} to
    turn off the interpretation of comments altogether.}

\item{stringsAsFactors}{logical: should character vectors be converted
    to factors?  Note that this is overridden by \code{as.is} and
    \code{colClasses}, both of which allow finer control.}

\item{na.strings}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after} 
    white space is stripped from the input, so \code{na.strings} 
    values may need their own white space stripped in advance.}

\item{...}{Arguments passed to \code{\link[utils:read.table]{utils::read.table()}}.}
}
\value{
A list.
}
\description{
These functions wrap the most common special cases of \code{\link[=list_any]{list_any()}}.
}
\examples{
(rds <- tor_example("rds"))
dir(rds)

list_rds(rds)

(tsv <- tor_example("tsv"))
dir(tsv)

list_tsv(tsv)

(mixed <- tor_example("mixed"))
dir(mixed)

list_rdata(mixed)

list_csv(mixed)

list_rdata(mixed, regexp = "[.]RData", ignore.case = FALSE)
}
\seealso{
Other general functions to import data: \code{\link{list_any}},
  \code{\link{load_csv}}, \code{\link{load_rds}}
}
\concept{general functions to import data}
