% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_with.R
\name{read_with}
\alias{read_with}
\title{Read multiple files into a list using any given reading function.}
\usage{
read_with(path = ".", .f, regexp = NULL, ignore.case = FALSE,
  invert = FALSE, ...)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{If \code{TRUE} return files which do \emph{not} match}

\item{...}{Additional arguments passed to \link{grep}.}
}
\value{
A list.
}
\description{
Read multiple files into a list using any given reading function.
}
\examples{
tor_example()

path <- tor_example("csv")
dir(path)

read_with(path, read.csv)

read_with(path, ~read.csv(.x, stringsAsFactors = FALSE))

(path_mixed <- tor_example("mixed"))
dir(path_mixed)

read_with(
  path_mixed, ~get(load(.x)),
  regexp = "[.]csv$",
  invert = TRUE
)

read_with(
  path_mixed, ~get(load(.x)),
  "[.]Rdata$",
  ignore.case = TRUE
)
}
\seealso{
Other general functions to import data: \code{\link{list_rds}},
  \code{\link{write_with}}
}
\concept{general functions to import data}
