% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readw-package.R
\docType{package}
\name{readw-package}
\alias{readw}
\alias{readw-package}
\title{readw: Read Multiple Files With Any Reading Function}
\description{
__readw__ allows you to read multiple files at once with any 
    reading function. Its main function, `readw::read_with()`, inputs a reader
    function and returns a modified version that reads any number of files and
    stores them into a list (provided the first argument of the reader function
    is a path to a file). Compared to other packages aimed to reading files, 
    __readw__ is more simpler, more flexible, and has less dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/readw}
  \item Report bugs at \url{https://github.com/maurolepore/readw/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
\keyword{internal}
