% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readw-package.R
\docType{package}
\name{readw-package}
\alias{readw}
\alias{readw-package}
\title{readw: Read Multiple Files With Any Reading Function}
\description{
Read multiple files with any reading function and
    store the output into a list. The main function, `read_with()`, is a
    function factory. It inputs any reader function and returns a
    modified, vectorized, version that reads multiple files at once.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/readw}
  \item Report bugs at \url{https://github.com/maurolepore/readw/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} [contractor]

}
\keyword{internal}
