% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwith-package.R
\docType{package}
\name{readwith-package}
\alias{readwith}
\alias{readwith-package}
\title{readwith: Read Multiple Files With Any Reading Function}
\description{
Use readwith to read, with your favorite reading function, 
    multiple files -- all at once. Its main function inputs a reader function 
    and returns a modified version that maps each files to one element of 
    a list (provided the first argument of the reader function is a path to a
    file). Compared to other packages aimed to reading files, readwith is more
    simpler, more flexible, and has less dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/readwith}
  \item Report bugs at \url{https://github.com/maurolepore/readwith/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
\keyword{internal}
