% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor-package.R
\docType{package}
\name{tor-package}
\alias{tor}
\alias{tor-package}
\title{tor: Read Multiple Files With Of Any Format From a Directory
    Into a List}
\description{
The goal of tor (to-R) is to help you to read
    multiple files from a single directory into R, and to do so as
    quickly, flexibly, and simply as possible. It makes a frequent task
    less painful. It has few dependencies and works well with the
    tidyverse (\url{https://www.tidyverse.org/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/tor}
  \item Report bugs at \url{https://github.com/maurolepore/tor/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
\keyword{internal}
