% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor-package.R
\docType{package}
\name{tor-package}
\alias{tor}
\alias{tor-package}
\title{tor: Import Multiple Files From a Single Directory at Once}
\description{
The goal of tor (to-R) is to help you to import multiple files from a single directory at once, and to do so as quickly, flexibly, and simply as possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/tor}
  \item \url{https://maurolepore.github.io/tor/}
  \item Report bugs at \url{https://github.com/maurolepore/tor/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
