% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_example.R
\name{tor_example}
\alias{tor_example}
\title{Easily access example data.}
\usage{
tor_example(path = NULL)
}
\arguments{
\item{path}{Length-1 character vector. A path to an available directory.}
}
\value{
A character string listing available directories or a length-1 string
giving a path to a directory.
}
\description{
Easily access example data.
}
\examples{
tor_example()

tor_example("csv")

dir(tor_example("csv"))
}
\author{
Copied from \code{readr_example()} from the \strong{readr} package, by Jim
Hester and colleagues.
}
\concept{helpers}
